requirements.txt
appium-python-client>=0.11
autoitpy==1.0.0
beautifulsoup4>=4.5.1
selenium>=3.0.2
xlrd>=0.9.3
PyYAML>=3.12
requests>=2.13.0
